# bank-service/src/main/java/org/application/user/entity/UserCardDto.java
package org.application.user.entity;

import lombok.*;

@AllArgsConstructor
@NoArgsConstructor
@Getter
@Setter
@EqualsAndHashCode
public class UserCardDto {
    private String name;

    private String surname;
}